AI Work + Models
Made Physically Morphed Dice with UV maps (Worked on the textures as well) AI Now Rolls It's Own Dice. Started Working on a sytem for the AI to be Able to move it's own cards. it recognized the cards it just need to activate a movement to the player in some way.